arreg;lando probelmas de logica

+++ waste-address-service/src/main/java/com/prueba/wasteservice/service/WasteAddressService.java
@@ -2,7 +2,10 @@ package com.prueba.wasteservice.service;
 
 import com.prueba.wasteservice.entity.WasteAddress;
 import com.prueba.wasteservice.repository.WasteAddressRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
@@ -14,48 +17,73 @@ import java.util.Objects;
 @Service
 public class WasteAddressService {
 
+    private static final Logger log = LoggerFactory.getLogger(WasteAddressService.class);
+
     @Autowired
     WasteAddressRepository wasteAddressRepository;
-
-    public List<WasteAddress> getAll() {
-        return wasteAddressRepository.findAll();
-    }
 
     public WasteAddress getWasteAddressById(Long id) {
         return wasteAddressRepository.findById(id).orElse(null);
     }
 
     public WasteAddress save(WasteAddress wasteAddress) {
-        return wasteAddressRepository.save(wasteAddress);
+        try {
+            WasteAddress savedWasteAddress = wasteAddressRepository.save(wasteAddress);
+
+            log.info("WasteManager saved successfully: {}", savedWasteAddress);
+            return savedWasteAddress;
+
+        }catch (DataAccessException e) {
+            log.error("Error saving WasteManager: {}", wasteAddress, e);
+            throw new RuntimeException("Failed to save WasteManager", e);
+        } catch (Exception e) {
+            log.error("Unexpected error when saving WasteManager: {}", wasteAddress, e);
+            throw e;
+        }
     }
 
-    public WasteAddress updateWasteAddress(Long id, WasteAddress wasteAddressDetails) throws Exception {
-        WasteAddress existingWasteAddress = wasteAddressRepository.findById(id).orElse(null);
+    public WasteAddress updateWasteAddress(Long id, WasteAddress wasteAddressDetails) {
+        try {
+            log.info("Attempting to update wasteAddress with id: {}", id);
+            WasteAddress wasteAddress = wasteAddressRepository.findById(id)
+                    .orElseThrow(() -> new RuntimeException("WasteAddress not found with id " + id));
 
-        boolean updateNeeded = false;
+            log.info("Found WasteAddress with id: {}, proceeding with update", id);
 
-        if (!Objects.equals(existingWasteAddress.getDireccion(), wasteAddressDetails.getDireccion())) {
-            existingWasteAddress.setDireccion(wasteAddressDetails.getDireccion());
-            updateNeeded = true;
-        }
+            if (wasteAddressDetails.getDireccion() != null && !wasteAddressDetails.getDireccion().isEmpty()) {
+                wasteAddress.setDireccion(wasteAddressDetails.getDireccion());
+            }
+            if (wasteAddressDetails.getIsEnabled() != null) {
+                wasteAddress.setIsEnabled(wasteAddressDetails.getIsEnabled());
+            }
+            if (wasteAddressDetails.getVersion() != null) {
+                wasteAddress.setVersion(wasteAddressDetails.getVersion());
+            }
 
-        if (!Objects.equals(existingWasteAddress.getIsEnabled(), wasteAddressDetails.getIsEnabled())) {
-            existingWasteAddress.setIsEnabled(wasteAddressDetails.getIsEnabled());
-            updateNeeded = true;
-        }
-
-        if (updateNeeded) {
-            existingWasteAddress.setLastModifiedDate(new Date());
-            return wasteAddressRepository.save(existingWasteAddress);
-        } else {
-            throw new Exception("No updates were performed as no changes were detected.");
+            WasteAddress updatedWasteAddress = wasteAddressRepository.save(wasteAddress);
+            log.info("wasteAddress with id: {} updated successfully", id);
+            return updatedWasteAddress;
+        } catch (Exception e) {
+            log.error("Error updating wasteAddress with id: {}, Error: {}", id, e.getMessage(), e);
+            throw new RuntimeException("Error updating wasteAddress with id " + id, e);
         }
     }
 
 
     public void deleteWasteAddress(Long id) {
-        WasteAddress wasteAddress = wasteAddressRepository.findById(id).orElse(null);
-        wasteAddressRepository.delete(wasteAddress);
+        try {
+            WasteAddress wasteAddress = wasteAddressRepository.findById(id).orElse(null);
+            if (wasteAddress == null) {
+                log.error("No WasteAddress found with id: {}", id);
+                throw new RuntimeException("No WasteAddress found with id " + id);
+            }
+
+            wasteAddressRepository.delete(wasteAddress);
+            log.info("WasteAddress with id: {} deleted successfully", id);
+        } catch (Exception e) {
+            log.error("Error deleting WasteAddress with id: {}, Error: {}", id, e.getMessage(), e);
+            throw new RuntimeException("Error deleting WasteAddress with id " + id, e);
+        }
     }
 
 }
